+++ apps/client/src/app/components/admin-market-data/create-asset-profile-dialog/interfaces/interfaces.ts
@@ -1,0 +1,7 @@
+import { AdminMarketDataItem } from '@ghostfolio/common/interfaces';
+
+export interface CreateAssetProfileDialogParams {
+  existingSymbols: AdminMarketDataItem[];
+  deviceType: string;
+  locale: string;
+}